+++ mayfly-sys/src/main/java/mayfly/sys/module/redis/controller/RedisController.java
@@ -36,7 +36,7 @@ public class RedisController {
 
     @MethodLog(level = MethodLog.LogLevel.DEBUG)
     @GetMapping("/{cluster}/{id}/scan")
-    public Result<?> scan(@PathVariable Boolean cluster, @PathVariable Integer id, @Valid ScanForm scanForm) {
+    public Result<?> scan(@PathVariable Boolean cluster, @PathVariable Long id, @Valid ScanForm scanForm) {
         RedisKeyCommands<String, byte[]> cmds = getKeyCmd(cluster, id);
         KeyScanVO scan = cluster ? KeyValueCommand.clusterScan(cmds, scanForm.getCount(), scanForm.getMatch())
                 : KeyValueCommand.scan(cmds, scanForm.getCursor(), scanForm.getCount(), scanForm.getMatch());
@@ -45,7 +45,7 @@ public class RedisController {
 
     @MethodLog(value = "查询redis value", resultLevel = MethodLog.LogLevel.DEBUG)
     @GetMapping("/{cluster}/{id}/value")
-    public Result<?> value(@PathVariable Boolean cluster, @PathVariable Integer id, String key) {
+    public Result<?> value(@PathVariable Boolean cluster, @PathVariable Long id, String key) {
         if (StringUtils.isEmpty(key)) {
             return Result.paramError("key不能为空!");
         }
@@ -54,7 +54,7 @@ public class RedisController {
 
     @MethodLog(value = "新增redis key value")
     @PostMapping("/{cluster}/{id}/value")
-    public Result<?> addKeyValue(@PathVariable Boolean cluster, @PathVariable Integer id, @Valid KeyValueForm keyValue) {
+    public Result<?> addKeyValue(@PathVariable Boolean cluster, @PathVariable Long id, @Valid KeyValueForm keyValue) {
         BaseRedisCommands<String, byte[]> cmds = cluster ? redisService.getClusterCmds(id) : redisService.getCmds(id);
         KeyValueCommand.addKeyValue(cmds, BeanUtils.copyProperties(keyValue, KeyInfo.class));
         return Result.success();
@@ -62,7 +62,7 @@ public class RedisController {
 
     @MethodLog("删除key")
     @DeleteMapping("/{cluster}/{id}")
-    public Result<?> delete(@PathVariable Boolean cluster, @PathVariable Integer id, String key) {
+    public Result<?> delete(@PathVariable Boolean cluster, @PathVariable Long id, String key) {
         if (StringUtils.isEmpty(key)) {
             return Result.paramError("key不能为空！");
         }
@@ -70,7 +70,7 @@ public class RedisController {
         return Result.success();
     }
 
-    private RedisKeyCommands<String, byte[]> getKeyCmd(Boolean cluster, Integer id) {
+    private RedisKeyCommands<String, byte[]> getKeyCmd(Boolean cluster, Long id) {
         return cluster ? redisService.getClusterCmds(id) : redisService.getCmds(id);
     }
 }
